tlsdate: add ppoll for arm64
Rust now calls `poll()` on init for all UNIX programs:
https://github.com/rust-lang/rust/commit/7d98d2207a20d95e731bfc742d89389a67d6618d

The README here says that tlsdate doesn't run under seccomp currently,
so this doesn't appear to be an _active_ fire. That said, a heuristic
led me to this file as something that needed updating, so here we go. :)

ARM32 and x86 both had `poll` in these files already. ARM64 doesn't have
the `poll` _syscall_; glibc appears to defer to `ppoll` for that on
systems which don't directly implement `poll`.

BUG=chromium:1192744
TEST=CQ

Change-Id: Ie2b7592ff54683e0f3a780a03ec8cc882c1e554c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/tlsdate/+/2862574
Reviewed-by: Mattias Nissler <mnissler@chromium.org>
Commit-Queue: George Burgess <gbiv@chromium.org>
Tested-by: George Burgess <gbiv@chromium.org>