tlsdate: Reset errno in validate_port()
tlsdate is failing with the "invalid port" error because the method
validating the port doesn't reset errno before trying to convert the
port value from string to integer. When errno is being set somewhere
else in the code, it causes the port validation to fail.

BUG=b:190205144
TEST=deployed change and ran tast test SystemProxyForSystemServices
which runs tlsdate behind a proxy

Change-Id: I2d75019924d13df5239d9506f974fe84541f1062
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/tlsdate/+/2945182
Reviewed-by: Jorge Lucangeli Obes <jorgelo@chromium.org>
Tested-by: Andreea-Elena Costinas <acostinas@google.com>
Commit-Queue: Andreea-Elena Costinas <acostinas@google.com>